Move SwbusConfig to swbus-config crate
