+++ crates/swbus-core/Cargo.toml
@@ -41,6 +41,7 @@ futures-core.workspace = true
 
 # Internal dependencies
 swbus-proto.workspace = true
+swbus-config.workspace = true
 
 [dev-dependencies]
 pretty_assertions.workspace = true
