+++ crates/swbus-core/src/mux/conn_store.rs
@@ -1,10 +1,10 @@
 use crate::mux::conn::SwbusConn;
-use crate::mux::swbus_config::{PeerConfig, RouteConfig};
 use crate::mux::SwbusConnInfo;
 use crate::mux::SwbusConnMode;
 use crate::mux::SwbusMultiplexer;
 use dashmap::{DashMap, DashSet};
 use std::sync::Arc;
+use swbus_config::{PeerConfig, RouteConfig};
 use tokio::task::JoinHandle;
 use tokio::time::Duration;
 use tracing::*;
