+++ crates/swbus-core/tests/common/test_executor.rs
@@ -4,8 +4,8 @@ use std::env;
 use std::fs::{self, File};
 use std::io::BufReader;
 use std::net::SocketAddr;
+use swbus_config::*;
 use swbus_core::mux::service::SwbusServiceHost;
-use swbus_core::mux::swbus_config::*;
 use swbus_edge::core_client::SwbusCoreClient;
 use swbus_proto::swbus::*;
 use tokio::sync::mpsc;
